Add LIBIPHO_SCREEN_FIFO to environment
The libipho-screen-server will be started as a deamon using
the startup scripts. This program needs to receive the
name of the fifo on the command line. Therefore, we need the
information about the name of the fifo already in the startup
scripts. Therefore, we set it in the environment for bash.